Add file_operations as `zea data`
